# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "CardIA API",
  "description": "Plataforma educativa de perfiles de tarjetas de crédito en México. No es asesoría financiera.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_v1_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Logout",
    "operationId": "logout_api_v1_auth_logout_post",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/cards": {
   "get": {
    "tags": [
     "cards"
    ],
    "summary": "List Cards",
    "operationId": "list_cards_api_v1_cards_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "institution",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Institution"
      }
     },
     {
      "name": "card_class",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Card Class"
      }
     },
     {
      "name": "benefit_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Benefit Type"
      }
     },
     {
      "name": "profile_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Profile Id"
      }
     },
     {
      "name": "max_annual_fee",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Annual Fee"
      }
     },
     {
      "name": "no_annual_fee",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "No Annual Fee"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "name",
        "annual_fee",
        "cat",
        "interest_rate",
        "credit_line_min",
        "benefits"
       ],
       "type": "string",
       "default": "name",
       "title": "Sort"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CardListResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/cards/facets": {
   "get": {
    "tags": [
     "cards"
    ],
    "summary": "Card Facets",
    "operationId": "card_facets_api_v1_cards_facets_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CardFacets"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/cards/{card_id}": {
   "get": {
    "tags": [
     "cards"
    ],
    "summary": "Card Detail",
    "operationId": "card_detail_api_v1_cards__card_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "card_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Card Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CardDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/chat/plan": {
   "post": {
    "tags": [
     "chat"
    ],
    "summary": "Plan",
    "description": "Paso 1 (D8): devuelve el plan explicito de tool calls antes de ejecutar.",
    "operationId": "plan_api_v1_chat_plan_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatPlanResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/chat/runs/{run_id}/execute": {
   "post": {
    "tags": [
     "chat"
    ],
    "summary": "Execute",
    "description": "Paso 2: ejecuta el plan por niveles y genera la respuesta.",
    "operationId": "execute_api_v1_chat_runs__run_id__execute_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatRunResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/chat/runs/{run_id}": {
   "get": {
    "tags": [
     "chat"
    ],
    "summary": "Get Run",
    "operationId": "get_run_api_v1_chat_runs__run_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatRunResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/chat/runs": {
   "get": {
    "tags": [
     "chat"
    ],
    "summary": "List Runs",
    "operationId": "list_runs_api_v1_chat_runs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "session_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 60
        },
        {
         "type": "null"
        }
       ],
       "title": "Session Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 30,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ChatRunResponse"
         },
         "title": "Response List Runs Api V1 Chat Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/overview": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Overview",
    "operationId": "overview_api_v1_admin_overview_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AdminOverview"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/admin/events": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Events",
    "operationId": "events_api_v1_admin_events_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "action",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 40
        },
        {
         "type": "null"
        }
       ],
       "title": "Action"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuditEventPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/chat-runs": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Chat Runs",
    "operationId": "chat_runs_api_v1_admin_chat_runs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ChatRunSummary"
         },
         "title": "Response Chat Runs Api V1 Admin Chat Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/users": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Users",
    "operationId": "users_api_v1_admin_users_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AdminUsers"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  }
 },
 "components": {
  "schemas": {
...
   "AdminOverview": {
    "properties": {
     "users": {
      "type": "integer",
      "title": "Users"
     },
     "cards": {
      "type": "integer",
      "title": "Cards"
     },
     "fees": {
      "type": "integer",
      "title": "Fees"
     },
     "benefits": {
      "type": "integer",
      "title": "Benefits"
     },
     "events": {
      "type": "integer",
      "title": "Events"
     },
     "chat_runs": {
      "type": "integer",
      "title": "Chat Runs"
     },
     "recommendations": {
      "type": "integer",
      "title": "Recommendations"
     },
     "by_action": {
      "items": {
       "$ref": "#/components/schemas/ActionCount"
      },
      "type": "array",
      "title": "By Action"
     },
     "llm_enabled": {
      "type": "boolean",
      "title": "Llm Enabled"
     },
     "database": {
      "type": "string",
      "title": "Database"
     },
     "data_generated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data Generated At"
     }
    },
    "type": "object",
    "required": [
     "users",
     "cards",
     "fees",
     "benefits",
     "events",
     "chat_runs",
     "recommendations",
     "by_action",
     "llm_enabled",
     "database",
     "data_generated_at"
    ],
    "title": "AdminOverview"
   },
   "AdminUsers": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/UserOut"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "items"
    ],
    "title": "AdminUsers"
   },
...
   "AuditEventPage": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/AuditEventOut"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "total",
     "items"
    ],
    "title": "AuditEventPage"
   },
...
   "CardDetail": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "institution": {
      "type": "string",
      "title": "Institution"
     },
     "card_class": {
      "type": "string",
      "title": "Card Class"
     },
     "cat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cat"
     },
     "annual_fee": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Annual Fee"
     },
     "interest_rate": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Interest Rate"
     },
     "credit_line_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credit Line Min"
     },
     "monthly_income_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Monthly Income Min"
     },
     "benefit_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Benefit Types"
     },
     "fee_counts": {
      "$ref": "#/components/schemas/FeeCounts"
     },
     "profile": {
      "$ref": "#/components/schemas/ProfileTag"
     },
     "requirements": {
      "$ref": "#/components/schemas/Requirements"
     },
     "fees": {
      "items": {
       "$ref": "#/components/schemas/FeeOut"
      },
      "type": "array",
      "title": "Fees"
     },
     "benefits": {
      "items": {
       "$ref": "#/components/schemas/BenefitOut"
      },
      "type": "array",
      "title": "Benefits"
     },
     "disclaimer": {
      "type": "string",
      "title": "Disclaimer"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "institution",
     "card_class",
     "cat",
     "annual_fee",
     "interest_rate",
     "credit_line_min",
     "monthly_income_min",
     "benefit_types",
     "fee_counts",
     "profile",
     "requirements",
     "fees",
     "benefits",
     "disclaimer"
    ],
    "title": "CardDetail"
   },
   "CardFacets": {
    "properties": {
     "institutions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Institutions"
     },
     "classes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Classes"
     },
     "benefit_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Benefit Types"
     },
     "profiles": {
      "items": {
       "$ref": "#/components/schemas/ProfileTag"
      },
      "type": "array",
      "title": "Profiles"
     },
     "annual_fee": {
      "$ref": "#/components/schemas/RangeOut"
     },
     "cat": {
      "$ref": "#/components/schemas/RangeOut"
     },
     "interest_rate": {
      "$ref": "#/components/schemas/RangeOut"
     }
    },
    "type": "object",
    "required": [
     "institutions",
     "classes",
     "benefit_types",
     "profiles",
     "annual_fee",
     "cat",
     "interest_rate"
    ],
    "title": "CardFacets"
   },
   "CardListResponse": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/CardSummary"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "total",
     "items"
    ],
    "title": "CardListResponse"
   },
   "CardSummary": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "institution": {
      "type": "string",
      "title": "Institution"
     },
     "card_class": {
      "type": "string",
      "title": "Card Class"
     },
     "cat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cat"
     },
     "annual_fee": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Annual Fee"
     },
     "interest_rate": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Interest Rate"
     },
     "credit_line_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credit Line Min"
     },
     "monthly_income_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Monthly Income Min"
     },
     "benefit_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Benefit Types"
     },
     "fee_counts": {
      "$ref": "#/components/schemas/FeeCounts"
     },
     "profile": {
      "$ref": "#/components/schemas/ProfileTag"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "institution",
     "card_class",
     "cat",
     "annual_fee",
     "interest_rate",
     "credit_line_min",
     "monthly_income_min",
     "benefit_types",
     "fee_counts",
     "profile"
    ],
    "title": "CardSummary"
   },
   "ChatPlanResponse": {
    "properties": {
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "session_id": {
      "type": "string",
      "title": "Session Id"
     },
     "plan": {
      "items": {
       "$ref": "#/components/schemas/PlanStep"
      },
      "type": "array",
      "title": "Plan"
     },
     "mode": {
      "type": "string",
      "enum": [
       "rules",
       "llm"
      ],
      "title": "Mode"
     }
    },
    "type": "object",
    "required": [
     "run_id",
     "session_id",
     "plan",
     "mode"
    ],
    "title": "ChatPlanResponse"
   },
   "ChatRequest": {
    "properties": {
     "message": {
      "type": "string",
      "maxLength": 1000,
      "minLength": 1,
      "title": "Message"
     },
     "session_id": {
      "type": "string",
      "maxLength": 60,
      "minLength": 1,
      "title": "Session Id"
     }
    },
    "type": "object",
    "required": [
     "message",
     "session_id"
    ],
    "title": "ChatRequest"
   },
   "ChatRunResponse": {
    "properties": {
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "session_id": {
      "type": "string",
      "title": "Session Id"
     },
     "question": {
      "type": "string",
      "title": "Question"
     },
     "plan": {
      "items": {
       "$ref": "#/components/schemas/PlanStep"
      },
      "type": "array",
      "title": "Plan"
     },
     "answer": {
      "type": "string",
      "title": "Answer"
     },
     "mode": {
      "type": "string",
      "enum": [
       "rules",
       "llm"
      ],
      "title": "Mode"
     },
     "status": {
      "type": "string",
      "enum": [
       "planned",
       "ok",
       "error"
      ],
      "title": "Status"
     },
     "cards": {
      "items": {
       "$ref": "#/components/schemas/CardSummary"
      },
      "type": "array",
      "title": "Cards"
     },
     "suggestions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Suggestions"
     },
     "disclaimer": {
      "type": "string",
      "title": "Disclaimer"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "duration_ms": {
      "type": "integer",
      "title": "Duration Ms"
     }
    },
    "type": "object",
    "required": [
     "run_id",
     "session_id",
     "question",
     "plan",
     "answer",
     "mode",
     "status",
     "cards",
     "suggestions",
     "disclaimer",
     "created_at",
     "duration_ms"
    ],
    "title": "ChatRunResponse"
   },
   "ChatRunSummary": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "user_email": {
      "type": "string",
      "title": "User Email"
     },
     "session_id": {
      "type": "string",
      "title": "Session Id"
     },
     "question": {
      "type": "string",
      "title": "Question"
     },
     "mode": {
      "type": "string",
      "title": "Mode"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "duration_ms": {
      "type": "integer",
      "title": "Duration Ms"
     },
     "steps": {
      "type": "integer",
      "title": "Steps"
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "user_email",
     "session_id",
     "question",
     "mode",
     "status",
     "duration_ms",
     "steps"
    ],
    "title": "ChatRunSummary"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginRequest": {
    "properties": {
     "email": {
      "type": "string",
      "maxLength": 200,
      "minLength": 3,
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest"
   },
...
   "TokenResponse": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "const": "bearer",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "user": {
      "$ref": "#/components/schemas/UserOut"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "TokenResponse"
   },
   "UserOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "role": {
      "type": "string",
      "enum": [
       "admin",
       "user"
      ],
      "title": "Role"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "created_at"
    ],
    "title": "UserOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}